feat: AppUserService 추상 클래스화 및 TxUserService 필드 타입 수정
