+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -10,7 +10,7 @@ import com.techcourse.config.DataSourceConfig;
 import com.techcourse.dao.UserDaoImpl;
 import com.techcourse.dao.UserHistoryDaoImpl;
 import com.techcourse.domain.User;
-import com.techcourse.service.AppUserService;
+import com.techcourse.service.AppUserServiceImpl;
 import com.techcourse.service.TxUserService;
 import com.techcourse.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
@@ -23,7 +23,7 @@ public class RegisterController {
 
     public RegisterController() {
         userService = new TxUserService(
-                new AppUserService(
+                new AppUserServiceImpl(
                         new UserDaoImpl(new JdbcTemplate(DataSourceConfig.getInstance())),
                         new UserHistoryDaoImpl(new JdbcTemplate(DataSourceConfig.getInstance()))
                 )
