+++ app/src/main/java/com/techcourse/service/AppUserService.java
@@ -1,42 +1,4 @@
 package com.techcourse.service;
 
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
-import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-
-public class AppUserService implements UserService {
-
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-
-    public AppUserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
-
-    @Override
-    public User findById(final long id) {
-        return userDao.findById(id)
-                .orElseThrow();
-    }
-
-    @Override
-    public User findByAccount(final String account) {
-        return userDao.findByAccount(account)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 계정입니다."));
-    }
-
-    @Override
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    @Override
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        User user = findById(id);
-        user.changePassword(newPassword);
-        userDao.update(user);
-        userHistoryDao.log(new UserHistory(user, createBy));
-    }
+public abstract class AppUserService implements UserService {
 }
